feat(backend): implement endpoint to change data use terms of sequence entries #763

+++ backend/src/main/kotlin/org/loculus/backend/utils/AccessionComparators.kt
@@ -7,13 +7,13 @@ typealias Version = Long
 
 object AccessionComparator : Comparator<Accession> {
     override fun compare(left: Accession, right: Accession): Int {
-        return left.toInt().compareTo(right.toInt())
+        return left.compareTo(right)
     }
 }
 
 object AccessionVersionComparator : Comparator<AccessionVersionInterface> {
     override fun compare(left: AccessionVersionInterface, right: AccessionVersionInterface): Int {
-        return when (val accessionResult = left.accession.toInt().compareTo(right.accession.toInt())) {
+        return when (val accessionResult = left.accession.compareTo(right.accession)) {
             0 -> left.version.compareTo(right.version)
             else -> accessionResult
         }
